Inicia desenvolvimento formulário Andamento/Acompanhamento do Projeto - Gestor de Projetos.

+++ src/main/java/br/com/live/sistema/body/BodyAtividadeProjeto.java
@@ -10,6 +10,6 @@ public class BodyAtividadeProjeto {
     public long idResponsavel;
     public String dataPrevInicio;
     public String dataPrevFim;
-    public int tempoPrevisto;
+    public double tempoPrevisto;
 
 }
